Remove InlineTaskClient, wrap code in FakeTaskClient
