+++ src/backend/common/deferred/queues/fake_queue.py
@@ -1,0 +1,11 @@
+from backend.common.deferred.queues.rq_queue import RQTaskQueue
+
+
+class FakeTaskQueue(RQTaskQueue):
+    """
+    A RQ-backed queue, but will run jobs inline
+    instead of making a HTTP request callback
+    """
+
+    def jobs(self):
+        return self._queue.jobs
